Fixed checks for files mimetype
All checks for file mimetype are failing because the magic.from_file
return a string instead of a binary string. Fixed that and print an
error message when sc-syncer can't tag or convert the file mentionning
the its mimetype.
So, sc-syncer was unable to tag or convert downloaded files for a
while..
This change should fix the issue #5 related to mp3 tagging from
sc-syncer.